ENH: change StdRegister to Register (depricate StdRegister)
